requirements.txt
aiohttp
fire
gin-config
lightning
loguru
opencv-python
pip-chill
scikit-learn
torchinfo
torchvision
wandb
